+++ tests/features/steps/auto_vertical.py
@@ -1,9 +1,8 @@
 # -*- coding: utf-8
-from __future__ import unicode_literals
+from __future__ import unicode_literals, print_function
+
 from textwrap import dedent
-
 from behave import then, when
-
 import wrappers