+++ tests/features/steps/basic_commands.py
@@ -4,13 +4,28 @@ Steps for behavioral style tests are defined in this module.
 Each step is defined by the string decorating it.
 This string is used to call the step in "*.feature" file.
 """
-from __future__ import unicode_literals
+from __future__ import unicode_literals, print_function
 
+import pexpect
+import subprocess
 import tempfile
 
-from behave import when
-import wrappers
+from behave import when, then
 from textwrap import dedent
+import wrappers
+
+
+@when('we list databases')
+def step_list_databases(context):
+    cmd = ['pgcli', '--list']
+    context.cmd_output = subprocess.check_output(cmd, cwd=context.package_root)
+
+
+@then('we see list of databases')
+def step_see_list_databases(context):
+    assert b'List of databases' in context.cmd_output
+    assert b'postgres' in context.cmd_output
+    context.cmd_output = None
 
 
 @when('we run dbcli')
@@ -28,7 +43,11 @@ def step_ctrl_d(context):
     """
     Send Ctrl + D to hopefully exit.
     """
+    # turn off pager before exiting
+    context.cli.sendline('\pset pager off')
+    wrappers.wait_prompt(context)
     context.cli.sendcontrol('d')
+    context.cli.expect_exact(pexpect.EOF, timeout=15)
     context.exit_sent = True
 
 
@@ -42,12 +61,12 @@ def step_send_help(context):
 
 @when(u'we send source command')
 def step_send_source_command(context):
-    with tempfile.NamedTemporaryFile() as f:
-        f.write(b'\?')
-        f.flush()
-        context.cli.sendline('\i {0}'.format(f.name))
-        wrappers.expect_exact(
-            context, context.conf['pager_boundary'] + '\r\n', timeout=5)
+    context.tmpfile_sql_help = tempfile.NamedTemporaryFile(prefix='pgcli_')
+    context.tmpfile_sql_help.write(b'\?')
+    context.tmpfile_sql_help.flush()
+    context.cli.sendline('\i {0}'.format(context.tmpfile_sql_help.name))
+    wrappers.expect_exact(
+        context, context.conf['pager_boundary'] + '\r\n', timeout=5)
 
 
 @when(u'we run query to check application_name')