+++ tests/test_prompt_utils.py
@@ -8,7 +8,6 @@ from pgcli.packages.prompt_utils import confirm_destructive_query
 
 def test_confirm_destructive_query_notty():
     stdin = click.get_text_stream('stdin')
-    assert stdin.isatty() is False
-
-    sql = 'drop database foo;'
-    assert confirm_destructive_query(sql) is None
+    if not stdin.isatty():
+        sql = 'drop database foo;'
+        assert confirm_destructive_query(sql) is None