+++ tests/test_smart_completion_multiple_schemata.py
@@ -1,4 +1,5 @@
-from __future__ import unicode_literals
+from __future__ import unicode_literals, print_function
+
 import itertools
 from metadata import (MetaData, alias, name_join, fk_join, join,
     schema, table, function, wildcard_expansion, column,