+++ changelog.rst
@@ -1,13 +1,39 @@
-1.10.4
-======
-
-Internal changes:
------------------
-* Use another psycopg2 connection for user query. (Thanks: `Matthieu Guilbert`_)
+Upcoming:
+=========
 
 Features:
 ---------
 * Add autocommit. (Thanks: `Matthieu Guilbert`_)
+
+Bug fixes:
+----------
+
+* Fix for loading/saving named queries from provided config file (#938). (Thanks: `Daniel Egger`_)
+* Set default port in `connect_uri` when none is given. (Thanks: `Daniel Egger`_)
+* Fix for error listing databases (#951). (Thanks: `Irina Truong`_)
+
+Internal:
+---------
+
+* Use another psycopg2 connection for user query. (Thanks: `Matthieu Guilbert`_)
+* Clean up and add behave logging. (Thanks: `Dick Marinus`_)
+* Require prompt_toolkit>=2.0.6. (Thanks: `Dick Marinus`_)
+* Improve development guide (Thanks: `Ignacio Campabadal`_)
+
+
+2.0.0:
+======
+
+* Update to ``prompt-toolkit`` 2.0. (Thanks: `Jonathan Slenders`_, `Dick Marinus`_, `Irina Truong`_)
+
+1.11.0
+======
+
+Features:
+---------
+
+* Respect `\pset pager on` and use pager when output is longer than terminal height (Thanks: `Max Rothman`_)
+
 
 1.10.3
 ======
@@ -879,3 +905,6 @@ Improvements:
 .. _`Saif Hakim`: https://github.com/saifelse
 .. _`Artur Balabanov`: https://github.com/arturbalabanov
 .. _`Kenny Do`: https://github.com/kennydo
+.. _`Max Rothman`: https://github.com/maxrothman
+.. _`Daniel Egger`: https://github.com/DanEEStar
+.. _`Ignacio Campabadal`: https://github.com/igncampa